+++ packages/frontend/next-app/src/features/login/ui/login-form.component.tsx
@@ -1,47 +1,13 @@
 "use client";
 
-import React, { useEffect, useState } from "react";
 import Link from "next/link";
-import { useForm } from "react-hook-form";
-import { zodResolver } from "@hookform/resolvers/zod";
-import type { LoginSchemaType } from "@/features/login/model/index";
-import { LoginSchema } from "@/features/login/model/index";
 import { Checkbox, Form, Input } from "@blog/components/src";
 import ErrorConfirm from "@/shared/ui/errorConfirmation/ErrorConfirm";
-import { signIn } from "next-auth/react";
-import { getEmailState, toggleEmailStorage } from "@/features/login/lib/email-storage";
+import { useLoginForm } from "@/features/login/hooks/useLoginForm";
 
 export default function LoginForm() {
-  const [rememberEmail, setRememberEmail] = useState(false);
-  const {
-    register,
-    handleSubmit,
-    formState: { errors },
-    setValue,
-  } = useForm<LoginSchemaType>({
-    resolver: zodResolver(LoginSchema),
-  });
-
-  const submitHandler = (form: LoginSchemaType) => {
-    toggleEmailStorage({ shouldRemember: rememberEmail, email: form.email });
-    signIn("credentials", {
-      ...form,
-      redirect: true,
-      callbackUrl: "/",
-    });
-  };
-
-  const toggleRememberEmail = () => {
-    setRememberEmail((prev) => !prev);
-  };
-
-  useEffect(() => {
-    const emailState = getEmailState();
-    if (emailState.shouldRemember) {
-      emailState.email && setValue("email", emailState.email);
-    }
-    setRememberEmail(emailState.shouldRemember);
-  }, [setValue]);
+  const { register, rememberEmail, toggleRememberEmail, handleSubmit, submitHandler, errors } =
+    useLoginForm();
 
   return (
     <section className="h-[450px] flex flex-col justify-evenly gap-y-5 border-2 px-4 rounded-2xl">
